create contract invoices
closes #4
