+++ app/Domain/Contract/Jobs/index.js
@@ -1,6 +1,6 @@
 const Queue = require('bull');
 const contractQueue = new Queue('create contract invoice', 'redis://127.0.0.1:6379');
-const createContractInvoice= use('App/Domain/Contract/Jobs/CreateContractInvoice'); 
+const contractJobs = use('App/Domain/Contract/Jobs/ContractJobs'); 
 
-contractQueue.process(createContractInvoice);
+contractQueue.process(contractJobs.createInvoices);
 module.exports = contractQueue;
