+++ package.json
@@ -32,6 +32,7 @@
     "@adonisjs/websocket": "^1.0.11",
     "adonisjs-querify": "^1.3.2",
     "bull": "^3.12.1",
+    "date-fns": "^2.8.1",
     "got": "^10.0.1",
     "mysql": "^2.16.0",
     "sphinx-promise": "^1.0.7",
